requirements.txt
pandas
numpy
scikit-learn
sentence-transformers
langchain
langchain-openai
langgraph
streamlit
python-dotenv
chromadb
